refactor(auth): merge storage_writer + storage_transaction into storage.py (ADR-0033 PR 1.1)
The `storage_state.json` persistence seam was one seam spelled as three files.
The split existed only to satisfy ADR-0008's 1000-line module cap — both dying
modules say so in their own docstrings — and it was re-joined at runtime by
function-local imports in BOTH directions across the storage <-> writer edge,
plus the template's lazy import back into the module it was extracted from.
Shallow modules with a lazy import per edge are the opposite of depth.

ADR-0033 lifts the cap for `_auth/` as a sanctioned-merge ceiling, which makes
this legal as a purely structural change. `storage.py` now carries, in labelled
sections mirroring the absorbed files: the lock primitives, the bounded acquire
+ secure-parent-dir prep, the transaction template and its three lock-failure
policies, the snapshot types, the CAS/merge math, the writer outcome types, and
the seven intent writers. `storage_writer.py` and `storage_transaction.py`
become re-export shims for one release, each carrying a removal note tied to the
next major.

This commit is a PURE MOVE. Every top-level definition was AST-compared against
its pre-merge source: 28 are byte-identical (docstrings aside) and the 10 that
differ differ ONLY by the cross-module re-joins this merge exists to remove —
four lazy imports deleted, `_storage.X` / `storage_writer.X` qualification
dropped to same-module lookups (still late-bound through the module globals, so
every existing patch seam keeps working), and the write bypass rebound from
`atomic_write_json` to `_write_state_unchecked`. That rename is ADR-0033
decision 2: the old alias collided with the name of the PUBLIC primitive the
guard protects, so gate and reader alike could not tell the guarded wrapper from
the bypass.

What deliberately did NOT move: `save_cookies_to_storage` stays in `storage.py`
under the same name (the ADR-0029-pinned monkeypatch seam that `_runtime/
lifecycle.py` late-binds and ~20 test files patch); the module-scope lock state
tests reset (`_FLOCK_UNAVAILABLE_WARNED`, the deadline constants) stays put;
`persist_minted_jar`'s #2108 ownership guard and write-ordering are bit-for-bit;
lock FILE paths and lock MECHANISMS are untouched; and the three writers still
hand-rolling the acquire stay unconverted (that is the next PR).
